miiverse-api: Introduce unified API error handlers
"Sending an error to the client" was previously an ad-hoc thing that varied dramatically between endpoints. Introduce a unified set of functions for it, add support to the logger.

Endpoints will need to be ported over to the new system - I've started with Communities for an example.

This commit changes client-visible error results, testing will be needed.