feat(web, console): Use IIFE bundle and unvendor PJAX
Since we have a bundler, might as well use it! Apparently previous work was (accidentally) using CJS format, which doesn't work on browser. Changing to IIFE means we don't get global scope anymore; which is needed for various places in the Juxt codebase where onclick= etc. are used.

Eventually need to move to addEventListener, but for now explicitly export things by setting them on window.

This allows us to stop vendoring PJAX and actually bundle it properly. Nice!